chore(fft): Delete the butterfly kernels' dead lint suppressions
Removing the six module-scoped allows across fixed.rs, hybrid.rs, and
stage.rs fires no many_single_char_names or too_many_arguments site:
the latter sits in the workspace's curated noise-allow set and no
butterfly identifier trips the former. The four per-site
wildcard_imports allows were dead likewise -- converted to expect, all
four reported unfulfilled -- and go with them. Also drops an
unnecessary usize cast in the gather reference test.

Item: ATLAS-APOLLO-BUTTERFLY-BLANKET-ALLOWS-2026-08-29

+++ crates/apollo-fft/src/application/execution/kernel/components/base128/tests.rs
@@ -653,7 +653,7 @@ where
     let mut reference = vec![T::from_precise(0.0); 2 * n];
     let bits = blocks.trailing_zeros();
     for b in 0..blocks {
-        let row = (b.reverse_bits() >> (usize::BITS - bits)) as usize;
+        let row = b.reverse_bits() >> (usize::BITS - bits);
         for j in 0..128 {
             reference[(row * 128 + j) * 2] = lanes[(j * blocks + b) * 2];
             reference[(row * 128 + j) * 2 + 1] = lanes[(j * blocks + b) * 2 + 1];
